tests/wpt/web-platform-tests/dom/nodes/Document-contentType/contentType/contenttype_datauri_01.html

<!DOCTYPE html>
<title>Data URI document.contentType === 'text/plain' when data URI MIME type is not set</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var iframe = document.createElement('iframe');
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "text/plain");
  }), false);
  iframe.src = "data:;,<!DOCTYPE html>";
  document.body.appendChild(iframe);
});
</script>
